+++ templates/jp-mining-note/base.js
@@ -209,18 +209,6 @@ function main() {
 {% endif %}
 {% endfor %}
 
-{% if EXTRA_JAVASCRIPT %}
-  try { // RUN: Extra Javascript
-    {% filter indent(width=4) %}
-    {{ EXTRA_JAVASCRIPT }}
-    {% endfilter %}
-  } catch (error) {
-    LOGGER.error("Error in extra javascript:");
-    LOGGER.errorStack(error.stack);
-  }
-{% endif %}
-
-
 }